requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.2
graphql-core==3.2.3
h11==0.14.0
httptools==0.6.1
idna==3.7
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.227.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
